requirements.txt
gradio==6.1.0
ollama==0.6.1
Pillow==12.0.0
Requests==2.32.5
torch==2.9.0
transformers==4.57.3
accelerate==1.12.0